Reservation details

+++ src/main/java/com/epf/rentmanager/dao/ClientDao.java
@@ -73,7 +73,7 @@ public class ClientDao {
         try {
             Connection connection = ConnectionManager.getConnection();
             PreparedStatement ps = connection.prepareStatement(FIND_CLIENT_QUERY);
-            ps.setLong(1, id);
+            ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
 
             if (rs.next()) {
